+++ src/main/java/cz/thepatik/welcomeplugin/WelcomePlugin.java
@@ -2,17 +2,18 @@ package cz.thepatik.welcomeplugin;
 
 import com.comphenix.protocol.ProtocolLibrary;
 import com.comphenix.protocol.ProtocolManager;
+import cz.thepatik.welcomeplugin.commands.CommandManager;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.io.File;
+import java.util.Objects;
 import java.util.logging.Logger;
 
 import static cz.thepatik.welcomeplugin.VersionCheck.*;
 import static java.lang.Double.parseDouble;
 
 public final class WelcomePlugin extends JavaPlugin {
-
     public static Logger logger;
     private ProtocolManager protocolManager;
 
@@ -32,42 +33,40 @@ public final class WelcomePlugin extends JavaPlugin {
     @Override
     public void onEnable() {
         // Plugin startup logic
-        //Check if config file exists
-        if (!fileExists(configFile())) {
-            getLogger().info("Creating config file...");
-            this.saveDefaultConfig();
-            if (fileExists(configFile())) {
-                getLogger().info("The config file successfully created!");
-                getLogger().info("The plugin is installed!");
-            } else {
-                getLogger().warning("There was a problem with creation of config file!");
-            }
+        this.saveDefaultConfig();
+
+        //Check ProtocolLibs
+        if (getServer().getPluginManager().getPlugin("ProtocolLib") == null) {
+            getLogger().warning("This plugin needs ProtocolLib in order to work!");
+            getLogger().warning("Disabling the plugin...");
+            Bukkit.getPluginManager().disablePlugin(this);
+            onDisable();
         }
-        //If config exists...
-          else {
-            //Check ProtocolLibs
-            if (getServer().getPluginManager().getPlugin("ProtocolLib") == null) {
-                getLogger().warning("This plugin needs ProtocolLib in order to work!");
-                getLogger().warning("Disabling the plugin...");
-                Bukkit.getPluginManager().disablePlugin(this);
-                onDisable();
-            }
-            //Check version
-            if (pluginVersion == parseDouble(getCurrentOnlineVersion())) {
-                getLogger().info("The plugin is up to date!");
-            } else {
-                getLogger().warning("There is a new version! Check GitHub!");
-                getLogger().info("Plugin version: " + pluginVersion + pluginVersionStage);
-                getLogger().info("Updated version: " + getCurrentOnlineVersion()+ pluginVersionStage);
-            }
-            protocolManager = ProtocolLibrary.getProtocolManager();
-            this.getConfig().set("pluginVersion", pluginVersion + pluginVersionStage);
+        //Check version
+        if (pluginVersion == parseDouble(getCurrentOnlineVersion())) {
+            getLogger().info("The plugin is up to date!");
+        } else {
+            getLogger().warning("There is a new version! Check GitHub!");
+            getLogger().info("Plugin version: " + pluginVersion + pluginVersionStage);
+            getLogger().info("Updated version: " + getCurrentOnlineVersion()+ pluginVersionStage);
+        }
+        //Register commands
+        getCommand("welcome").setExecutor(new CommandManager());
+
+        //Register ProtocolLib
+        protocolManager = ProtocolLibrary.getProtocolManager();
+
+        String pluginVersionConfig = pluginVersion + pluginVersionStage;
+
+        //Set version in config
+        if (!pluginVersionConfig.equals(this.getConfig().getString("plugin-version"))) {
+            this.getConfig().set("plugin-version", pluginVersion + pluginVersionStage);
             this.saveConfig();
-            //Finally the plugin is loaded...
-            getLogger().info("The plugin is loaded!");
-            getServer().getPluginManager().registerEvents(new PlayerListener(), this);
         }
 
+        //Finally the plugin is loaded...
+        getLogger().info("The plugin is loaded!");
+        getServer().getPluginManager().registerEvents(new PlayerListener(this), this);
     }
 
     @Override