+++ CHANGELOG.md
@@ -1,16 +1,27 @@
 # Changelog
 
+## 0.4-alpha
+* Added CommandManager
+* New commands added
+  * /welcome help - Shows list of commands
+  * /welcome update - Checks for update
+  * /welcome version - Shows installed version of plugin
+  * ~~/welcome showcreditsto [who] - Change option to who show credits (Now only for everyone or newcomers)~~
+* ~~Added option in config to display credits(or whatever you want) to newcomers~~
+* Added permissions to commands
+
+
 ## 0.3-alpha
-* Add onJoinEvent - when player joins first time, show credits
+* Added onJoinEvent - when player joins first time, show credits
 * Cleared code from junk
 * Started to work on display message when player joins
 
 ## 0.2-alpha
-* Add ProtocolLib dependency
-* Add config
+* Added ProtocolLib dependency
+* Added config
 * Started to work on onJoinEvent
 
 ## 0.1-alpha
 * Hello World!
 * Started to write code
-* Add versioning
+* Added versioning